Refactor Utility.hh header into cohesive units
Move each class/function into dedicated header to clarify use and
simplify future updates.
