+++ examples/basic/tasking.cc
@@ -67,11 +67,9 @@ execute_cpu_iterations(int64_t num_iter, TaskGroup_t* task_group, int64_t n,
 //============================================================================//
 
 int
-main(int argc, char** argv)
+main()
 {
     _pause_collection;  // for VTune
-
-    ConsumeParameters(argc, argv);
 
     auto hwthreads      = std::thread::hardware_concurrency();
     auto default_fib    = 30;
