+++ source/TaskRunManager.cc
@@ -21,10 +21,10 @@
 #include "PTL/TaskRunManager.hh"
 
 #include "PTL/Config.hh"
+#include "PTL/GetEnv.hh"
 #include "PTL/TaskManager.hh"
 #include "PTL/ThreadPool.hh"
 #include "PTL/Threading.hh"
-#include "PTL/Utility.hh"
 
 #include <iostream>
