+++ source/VUserTaskQueue.cc
@@ -27,7 +27,6 @@
 
 #include "PTL/VUserTaskQueue.hh"
 #include "PTL/TaskRunManager.hh"
-#include "PTL/Utility.hh"  // for PTL
 #include <cstdint>         // for intmax_t
 #include <thread>          // for thread
